backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt[token_blacklist]>=5.3
django-environ>=0.11
django-cors-headers>=4.3
django-filter>=23.5
django-csp>=3.8
drf-spectacular>=0.27
dj-database-url>=2.1
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.6
cloudinary>=1.40
celery>=5.3
redis>=5.0
sentry-sdk>=2.0
mercadopago>=2.2
sib-api-v3-sdk>=7.6
Pillow>=10.2
